Fixed broken link
It turns out that if you upload an image with the exact same file name as an image you just deleted off GitHub, GitHub will erroneously recover and reupload the image you just deleted, instead of uploading the new image you attempted to upload. To get around this I added an underscore after the "8" in my new image's name. However, this broke the link in the README, so this update added the underscore to fix it.